// src/ts/column-type.ts
// automatically generated by the FlatBuffers compiler, do not modify

/**
 * @enum {number}
 */
export enum ColumnType{
  Byte= 0,
  UByte= 1,
  Bool= 2,
  Short= 3,
  UShort= 4,
  Int= 5,
  UInt= 6,
  Long= 7,
  ULong= 8,
  Float= 9,
  Double= 10,
  String= 11,
  Json= 12,
  DateTime= 13,
  Binary= 14
}